Move progress_sync to util.hpp
